edit code

+++ app/src/main/java/defpackage/btj.java
@@ -1,7 +1,7 @@
 package defpackage;
 
-/* renamed from: btj  reason: default package */
+/* renamed from: btj reason: default package */
 /* compiled from: PG */
 public final class btj {
-    public static final Long a = 86400000L;
+    public static final java.lang.Long a = java.lang.Long.valueOf(86400000);
 }
